requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
decorator==4.4.2
ffmpeg-python==0.2.0
filelock==3.9.0
Flask==2.2.2
Flask-WTF==1.1.1
future==0.18.3
huggingface-hub==0.11.1
idna==3.4
imageio==2.25.0
imageio-ffmpeg==0.4.8
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
more-itertools==9.0.0
moviepy==1.0.3
numpy==1.24.1
#nvidia-cublas-cu11==11.10.3.66
#nvidia-cuda-nvrtc-cu11==11.7.99
#nvidia-cuda-runtime-cu11==11.7.99
#nvidia-cudnn-cu11==8.5.0.96
openai-whisper==20230117
packaging==23.0
Pillow==9.4.0
proglog==0.1.10
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
semantic-version==2.10.0
setuptools-rust==1.5.2
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
WTForms==3.0.1
